BC-11983 - Move oauth-adapter module from modules to infra
