+++ apps/server/src/infra/oauth-adapter/dto/client-credentials-grant-token-request.ts
