+++ apps/server/src/infra/oauth-adapter/dto/oauth-token.response.ts
