+++ apps/server/src/infra/oauth-adapter/loggable/token-request-loggable-exception.spec.ts
