+++ apps/server/src/infra/schulconnex-client/schulconnex-rest-client.spec.ts
@@ -1,6 +1,6 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { Logger } from '@infra/logger';
-import { OauthAdapterService, OAuthTokenDto } from '@modules/oauth-adapter';
+import { OauthAdapterService, OAuthTokenDto } from '@infra/oauth-adapter';
 import { HttpService } from '@nestjs/axios';
 import { axiosResponseFactory } from '@testing/factory/axios-response.factory';
 import { of } from 'rxjs';
