+++ apps/server/src/infra/tsp-client/tsp-client-factory.ts
@@ -1,5 +1,5 @@
 import { AxiosErrorLoggable } from '@core/error/loggable';
-import { ClientCredentialsGrantTokenRequest, OauthAdapterService, OAuthGrantType } from '@modules/oauth-adapter';
+import { ClientCredentialsGrantTokenRequest, OauthAdapterService, OAuthGrantType } from '@infra/oauth-adapter';
 import { Inject, Injectable } from '@nestjs/common';
 import { AxiosError } from 'axios';
 import * as jwt from 'jsonwebtoken';
